+++ extension/jest.config.js
@@ -2,14 +2,14 @@
 
 module.exports = {
   coverageDirectory: 'coverage/jest',
-  coveragePathIgnorePatterns: ['<rootDir>/src/test/'],
+  coveragePathIgnorePatterns: ['<rootDir>/src/test/', '/node_modules/'],
   coverageReporters: ['json'],
   testEnvironment: 'node',
   testPathIgnorePatterns: [
     '<rootDir>/src/test/',
     '<rootDir>/dist/',
-    '<rootDir>/.vscode-test',
-    '<rootDir>/.wdio-vscode-service'
+    '<rootDir>/.vscode-test/',
+    '<rootDir>/.wdio-vscode-service/'
   ],
   transform: {
     '^.+\\.(t|j)sx?$': ['@swc/jest']