+++ extension/src/test/suite/experiments/index.test.ts
@@ -63,7 +63,7 @@ import * as Telemetry from '../../../telemetry'
 import { EventName } from '../../../telemetry/constants'
 import * as VscodeContext from '../../../vscode/context'
 import { Title } from '../../../vscode/title'
-import { EXP_RWLOCK_FILE, ExperimentFlag } from '../../../cli/dvc/constants'
+import { ExperimentFlag } from '../../../cli/dvc/constants'
 import { DvcExecutor } from '../../../cli/dvc/executor'
 import { WorkspacePlots } from '../../../plots/workspace'
 import {
@@ -1245,32 +1245,29 @@ suite('Experiments Test Suite', () => {
 
       const webview = await experiments.showWebview()
       const mockMessageReceived = getMessageReceivedEmitter(webview)
-      const mockExperimentIds = ['exp-e7a67', 'test-branch']
+      const mockExperimentIds = ['exp-e7a67', 'exp-83425']
 
       stubWorkspaceExperimentsGetters(dvcDemoPath, experiments)
       const mockPid = 1234
-      const mockGetPidFromFile = stub(FileSystem, 'getPidFromFile').resolves(
-        mockPid
-      )
+      const mockGetPidFromFile = stub(FileSystem, 'getPidFromFile')
+        .onFirstCall()
+        .resolves(mockPid)
+        .onSecondCall()
+        .resolves(undefined)
       const mockProcessExists = stub(
         ProcessExecution,
         'processExists'
       ).resolves(true)
 
       mockMessageReceived.fire({
-        payload: [
-          ...mockExperimentIds.map(id => ({ executor: Executor.DVC_TASK, id })),
-          { executor: Executor.WORKSPACE, id: EXPERIMENT_WORKSPACE_ID }
-        ],
-        type: MessageFromWebviewType.STOP_EXPERIMENT
+        payload: mockExperimentIds,
+        type: MessageFromWebviewType.STOP_EXPERIMENTS
       })
 
       await Promise.all([experimentsKilled, workspaceStopped])
 
-      expect(mockQueueKill).to.be.calledWith(dvcDemoPath, ...mockExperimentIds)
-      expect(mockGetPidFromFile).to.be.calledWithExactly(
-        join(dvcDemoPath, EXP_RWLOCK_FILE)
-      )
+      expect(mockQueueKill).to.be.calledWith(dvcDemoPath, 'exp-e7a67')
+      expect(mockGetPidFromFile).to.be.calledTwice
       expect(mockProcessExists).to.be.calledWithExactly(mockPid)
       expect(mockStopProcesses).to.be.calledWithExactly([mockPid])
     }).timeout(WEBVIEW_TEST_TIMEOUT)